feat: re-organize context for templates using a mixin
context that will be shared across views of the districts app is
moved into the DistrictContextMixin. This reduces context duplication,
both when different views need the same context and also when the view
may not need the context, but the template it extends does need it.
Note that because of how the MRO and mixins work, DistrictContextMixin
inherits from TemplateView.