feat: highlight the active district in the sidebar
moved the sidebar nav to a separate template for modularity. Also
note that when the sidebar is rendered from the IndexView, the
template engine will silently treat current_district as None since
it is not passed in the context. This is fine since the sidebar
highlight will not be rendered in this case.