chore: set up streamlit service for use in AWS
since the web and streamlit services run on the same load balancer
in AWS, the http path defined in the manifest of each service must
be different from each other. the http path of the streamlit service
(streamlit_service) is also passed to the streamlit run command in
the entrypoint.sh script. also, the healthcheck path must be
defined in the manifest so that the service does not
continually restart.