feat(compose): ensure postgres server is ready
introduce a healthcheck for the postgres service to ensure the
web and pgweb services do not start until the postgres server is ready.
this ensures we don't get an Admin connection to PostgreSQL fail
for web and an Error: connection refused error for pgweb when there
is no postgres service already running when web or pgweb are started.
This usually happens when the containers are first created.