fix(infra): use GitHub deployment environments feature
instead of using the COPILOT_ENVIRONMENT variable, we now use
the GitHub environment feature, which is set based on the branch
or tag being deployed.
Note that because this workfow is triggered by a push to main,
we can safely assume that the desired environment is dev. Running this
workflow manually (workflow_dispatch) from a branch that is not main
would potentially also deploy to dev, which is desired behavior. Only
tags are intended to be used to deploy to test or prod environments.